backend/pyproject.toml
[tool.poetry]
authors = ["Riley Ashton <37090587+riley-ashton@users.noreply.github.com>"]
description = "Demo project for fastapi and sqlmodel full stack"
name = "fastapi-sqlmodel"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
alembic = "^1.10.2"
arq = "^0.25.0"
asyncpg = "^0.27.0"
fastapi = {extras = ["all"], version = "^0.95.0"}
httpx = "^0.23.3"
passlib = {extras = ["argon2"], version = "^1.7.4"}
python = "^3.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
sqlalchemy-utils = "^0.40.0"
sqlmodel = "^0.0.8"
uvicorn = "^0.21.1"

[tool.poetry.group.dev.dependencies]
aiosqlite = "^0.18.0"
bandit = "^1.7.4"
black = "^23.1.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
mypy = "^1.0.1"
pydeps = "^1.11.1"
pytest = "^7.2.1"
pytest-asyncio = "^0.21.0"
safety = "^2.3.5"

[tool.isort]
profile = "black"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
